+++ csharp/tools/gendarme-runner/src/main/java/org/sonar/dotnet/tools/gendarme/GendarmeRunner.java
@@ -28,7 +28,6 @@ import org.apache.commons.io.FileUtils;
 import org.apache.commons.lang.StringUtils;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import org.sonar.api.utils.command.Command;
 import org.sonar.api.utils.command.CommandExecutor;
 import org.sonar.dotnet.tools.commons.utils.ZipUtils;
 import org.sonar.plugins.csharp.api.visualstudio.VisualStudioProject;
@@ -41,7 +40,7 @@ public class GendarmeRunner { // NOSONAR : can't mock it otherwise
 
   private static final Logger LOG = LoggerFactory.getLogger(GendarmeRunner.class);
 
-  private static final int MINUTES_TO_MILLISECONDS = 60000;
+  private static final long MINUTES_TO_MILLISECONDS = 60000;
   private static final String EMBEDDED_VERSION = "2.10";
 
   private File gendarmeExecutable;
@@ -83,8 +82,8 @@ public class GendarmeRunner { // NOSONAR : can't mock it otherwise
   }
 
   /**
-   * Creates a pre-configured {@link GendarmeCommandBuilder} that needs to be completed before running the {@link #execute(Command, int)}
-   * method.
+   * Creates a pre-configured {@link GendarmeCommandBuilder} that needs to be completed before running the
+   * {@link #execute(GendarmeCommandBuilder, int)} method.
    * 
    * @param solution
    *          the solution to analyse
@@ -97,8 +96,8 @@ public class GendarmeRunner { // NOSONAR : can't mock it otherwise
   }
 
   /**
-   * Creates a pre-configured {@link GendarmeCommandBuilder} that needs to be completed before running the {@link #execute(Command, int)}
-   * method.
+   * Creates a pre-configured {@link GendarmeCommandBuilder} that needs to be completed before running the
+   * {@link #execute(GendarmeCommandBuilder, int)} method.
    * 
    * @param project
    *          the VS project to analyse
